First editor experiment
Add some files from a first experiment in the hopes of finding a good
scene editor.

The idea is that a scene can be created in Tiled Editor, which can then
be exported in the form of a JSON. This could either be used directly by
the game, or further processed by some script.