Added 3 takes for components. Implemented components for pos & sprites.
Alpha is an approach that allows for static typing for dependencies,
giving compile-type errors at failure.
But their addition is not enforced by anything...

Beta simply has the components stored in a field, and dependencies are
given as non-null arguments to the constructors.
As simple as it gets.

Gamma has an object explicitly for storing components, which checks
dependencies and calls Init for all.
Quite similar to the prior, but with an intermediate handling some stuff