Created a Utils class and made Player not perform its own LoadContent.
The NotNull methods are now in an Utils class, and the NotNullDependency
method now has a custom exception related to it.

Players are given a Texture2D now,
instead of being called by LoadContent to load content.
Needs to be improved, as it should preferably be passed it in the
constructor.

Also cleaned up some TODO comment's, and made sprite print to std-error
if drawn with no texture.