Made the input system more complete, and made stuff use it.
The input system now handles mouse too, and was refactored to not use an
enum, as disagreement ocurred with using bitmasks, and without that an
enum will become a hassle to compare/mutate in various ways.
The class for ProperKeyState uses several properties to make checking
some things rather simple, aliasing otherwise common conditionals.

The dict of prior state was made to delete keys once they are Up,
to ensure it doesn't slowly get larger with every new type of keypress,
as it was considered superfluous and easy to do.
Also simplifies some logic using it.

To handle mouse, the keyboard- and mouse-keys had to be joined somehow,
which was done by implementing an Either<L,R> class.
The mouse-keys only existed as properties of MouseState,
so to better fascilitate using `Either`, an enum was made for them.