A somewhat too big commit. Makes input bindable actions + more.
Made Either<> more tolerant to same type (still not reccomended)
and gave it implicit casts.

Moved Input related files into their own folder.

Gave IGameState their own Input property, allowing for input to not be
broken by changing current state and returning after changing input.
This is to ie. avoid actions that depend on both edgeup and edgedown
from getting locked (until key is pressed again) during such a
state-transition.
Think +forward in source, but no -forward was caught.
In other words, a state's personal input is paused alongside the state.

Split Input into two classes, one for the underlying state-dict code
(now called StateMap), and another to poll input and merge the keys.
It also keeps responsibility for what doesn't act as a signalgraph,
such as mouse-position and scrolling.
The fullscreen-fix was moved into its own static class, in case anything
else would want to use it.

Created an Actions enum, which are the actions relevant to the game.
Can be considered somewhat like a concmd (except doesn't trigger a hook)
or a key in-and-of-itself.
The idea is that this enum defines the Actions a user can bind in their
config file, including things such as hotkeys like togglefullscreen and
interaction such as LeftMouse (for moving) and menu-navigation.

A Hotkeys class (namechange pending) was created.
It ties together regular input and the bindable Actions that logic
checks when handling ui.
Uses the StateMap like Input, so the Actions can be considered something
of a (custom) logical keyboard that normal input has been translated to.
It is currently also where the Actions are bound to key-combinations,
though this is slated for change.
The same file currently holds the Actions enum itself.

Created a HandleInput method that is meant to run pre-Update,
so while Update also allows for handling input, there is a way to ensure
it happens before other objects Update, and some compartmentalization.

(Re)Moved the ProperKeyState struct/class, and renamed it to SignalState
as it better describes what it is (the state of a discrete signal-graph)
although worse at what it is intended to be used for.

Moved the IsActive/focus check out of Input so it doesn't need a ref to
the Game instance. So together with the fullscreen-fix being moved,
neither Game nor Screen is needed in Input.

Did however forget to add the check to the states, need to do that.
Also needs to figure out a good way to merge the Actions between
the states Input and the global one.
Could remove the global one entirely, but stuff like Screen uses it,
which is why it is considered an Input for global hotkeys...