Cleaned up imports etc. and fixed code-duplciation in from gamestates.
Updating the gamestate's input is now done in main, so each gamestate
does not have to implement it themselves. Might be argued a state wants
to skip it, but I really doubt that.
This also fixes last commits issue that only Main's input-instance
checked for focus.

Removed HandleInput from gamestates, it is now instead only in stuff
implementing the interface, which are currently only called by Playing
calling Scene (now an IInputHandle) in its Update-hook, which then
goes thorugh its list of inputhandlers to handle.

Added the list to the scene-parsing code too, so it can add stuff to it.

Renamed the functions in StateMap and its subclasses. Rather than
HandleInput/Keys, the verb Update is now preferred.
Also cleaned some method-signatures there.

Cleaned some unused imports and old comments / commented-out code.

Commented out IsAltEnterPressed() from Screen, and the variable it used
to keep track of prior frames state of the hotkey.
Used the action properly now instead.
Forgot to clean out/remove the commented code...