Add speech functionality (#79)
The character can now say things. The message will be displayed for a
time depending on the length of the string. If Say() is called several
times, the strings will be queued up and done once each.

I think we need a smaller font. I used a line width similar to Monkey
Island.